+++ demo/examples.js
@@ -261,7 +261,8 @@ examples_main = {
       { name: "docu1", file: "efficiency.root", item: "eff1", title: "TEfficiency example from documentation" },
       { name: "docu2", file: "efficiency.root", item: "eff2", title: "TEfficiency example from documentation" },
       { name: "docu3", file: "efficiency.root", item: "eff3", title: "using TMath::BetaDist" },
-      { name: "docu4", file: "efficiency.root", item: "eff4", title: "drawing functions from TEfficiency" }
+      { name: "docu4", file: "efficiency.root", item: "eff4", title: "drawing functions from TEfficiency" },
+      { name: "2d", file: "efficiency.root", item: "eff5", title: "drawing 2-D efficiency" }
    ],
    TLatex : [
      { name: "latex", file: "latex.root", layout: "grid2x2", items: ['lva','ex1','ex2','ex3'], title: "Internal TLatex processor" },